Timeline responsive handling
Edgecase not conisidered for things like when the plips stack up, the progress bar
is not able to detect when items are stacked as it doesnt have an id to anchor to.
Had to settle on a solution where it is only visible above 1200px container width
as a pragmatic approach thinking its unlikely to cause many issues above 1200px
until I look at rebuilding. The obviouss soilution is some sort if JS approach, but that can come
later down the line as the data wont change